+++ examples/udf/smooth_savitzky_golay.py
@@ -1,15 +1,18 @@
-from typing import Dict
 import xarray
 from scipy.signal import savgol_filter
-from openeo_udf.api.datacube import DataCube
 
-def apply_datacube(cube: DataCube, context: Dict) -> DataCube:
+from openeo.udf import XarrayDataCube
+
+
+def apply_datacube(cube: XarrayDataCube, context: dict) -> XarrayDataCube:
     """
-    Applies a savitzky-golay smoothing to a timeseries datacube.
-    This UDF preserves dimensionality, and assumes a datacube with a temporal dimension 't' as input.
+    Apply Savitzky-Golay smoothing to a timeseries datacube.
+    This UDF preserves dimensionality, and assumes an input
+    datacube with a temporal dimension 't' as input.
     """
     array: xarray.DataArray = cube.get_array()
     filled = array.interpolate_na(dim='t')
     smoothed_array = savgol_filter(filled.values, 5, 2, axis=0)
-    return DataCube(xarray.DataArray(smoothed_array,dims=array.dims,coords=array.coords))
-
+    return XarrayDataCube(
+        array=xarray.DataArray(smoothed_array, dims=array.dims, coords=array.coords)
+    )